requirements.txt
absl-py==1.0.0
anyio==3.5.0
asgiref==3.5.0
astunparse==1.6.3
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Django==4.0.3
fastapi==0.75.0
flatbuffers==2.0
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h11==0.13.0
h5py==3.6.0
idna==3.3
iniconfig==1.1.1
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
Markdown==3.3.6
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
Pillow==9.0.1
pluggy==1.0.0
protobuf==3.19.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
pyparsing==3.0.7
pytest==7.1.0
python-multipart==0.0.5
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.17.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.6
Werkzeug==2.0.3
wrapt==1.14.0
